+++ sentry-spring/src/test/kotlin/io/sentry/spring/tracing/SentryTransactionAdviceTest.kt
@@ -100,7 +100,8 @@ class SentryTransactionAdviceTest {
 
     @Test
     fun `when transaction is already active, does not start new transaction`() {
-        whenever(hub.span).thenReturn(SentryTracer(TransactionContext("aTransaction", "op"), hub))
+        whenever(hub.options).thenReturn(SentryOptions())
+        whenever(hub.span).then { SentryTracer(TransactionContext("aTransaction", "op"), hub) }
 
         sampleService.methodWithTransactionNameSet()