+++ sentry/src/test/java/io/sentry/instrumentation/file/SentryFileWriterTest.kt
@@ -16,13 +16,14 @@ import kotlin.test.assertEquals
 class SentryFileWriterTest {
     class Fixture {
         val hub = mock<IHub>()
-        val sentryTracer = SentryTracer(TransactionContext("name", "op"), hub)
+        lateinit var sentryTracer: SentryTracer
 
         internal fun getSut(
             tmpFile: File,
             activeTransaction: Boolean = true,
         ): SentryFileWriter {
             whenever(hub.options).thenReturn(SentryOptions())
+            sentryTracer = SentryTracer(TransactionContext("name", "op"), hub)
             if (activeTransaction) {
                 whenever(hub.span).thenReturn(sentryTracer)
             }